Crate aws_types

Cross-service types for the AWS SDK.

Re-exports

pub use sdk_config::SdkConfig;

Modules

app_name

New-type for a configurable app name.

build_metadata

Metadata type used by all service SDKs to provide information about the environment the SDK was compiled for.

endpoint_config

Newtypes for endpoint-related parameters

os_shim_internal

Abstractions for testing code that interacts with the operating system:

region

Region type for determining the endpoint to send requests to.

request_id

AWS-specific request ID support

sdk_config

AWS Shared Config

Macros

docs_for

A macro that generates docs for selected fields of SdkConfig.

Structs

SigningName

The name of the service used to sign this request