aws_types

Module endpoint_config

Newtypes for endpoint-related parameters

Parameters require newtypes so they have distinct types when stored in layers in config bag.

Structs

AccountIdEndpointModeParseError

Error encountered when failing to parse a string into AccountIdEndpointMode.

EndpointUrl

Newtype for endpoint_url

UseDualStack

Newtype for use_dual_stack

UseFips

Newtype for use_fips

Enums

AccountIdEndpointMode

Setting to control the account ID-based routing behavior.