aws_types
Module sdk_config

AWS Shared Config

This module contains a shared configuration representation that is agnostic from a specific service.

Modules

unified_docs

Unified docstrings to keep crates in sync. Not intended for public use

Structs

Builder

Builder for AWS Shared Configuration

RetryConfig

Retry configuration for requests.

SdkConfig

AWS Shared Configuration

SharedAsyncSleep

Wrapper type for sharable AsyncSleep

SharedCredentialsProvider

Credentials Provider wrapper that may be shared

SharedHttpClient

Shared HTTP client for use across multiple clients and requests.

SharedTimeSource

Time source structure used inside SDK

StalledStreamProtectionConfig

Configuration for stalled stream protection.

TimeoutConfig

Top-level configuration for timeouts

Enums

RequestChecksumCalculation

Determines when a checksum will be calculated for request payloads. Values are:

ResponseChecksumValidation

Determines when checksum validation will be performed on response payloads. Values are:

Traits

TimeSource

Trait with a now() function returning the current time