aws_types

Module service_config

Code for extracting service config from the user’s environment.

Modules

builder

Builder for ServiceConfigKey.

Structs

ServiceConfigKey

A struct used with the LoadServiceConfig trait to extract service config from the user’s environment.

Traits

LoadServiceConfig

Implementers of this trait can provide service config defined in a user’s environment.