Crate aws_types

Cross-service types for the AWS SDK.

Re-exports

pub use sdk_config::SdkConfig;

Modules

app_name

New-type for a configurable app name.

build_metadata

Metadata type used by all service SDKs to provide information about the environment the SDK was compiled for.

configDeprecated

AWS Shared Config (deprecated, replaced with sdk_config)

credentials

AWS SDK Credentials

endpoint

AWS SDK endpoint support.

http_connector

Default connectors based on what TLS features are active. Also contains HTTP-related abstractions that enable passing HTTP connectors around.

region

Region type for determining the endpoint to send requests to.

sdk_config

AWS Shared Config

Structs

Credentials

AWS SDK Credentials

SigningService

The name of the service used to sign this request